+++ ntex-error/src/repr.rs
@@ -1,6 +1,6 @@
 use std::{error, fmt, panic::Location, sync::Arc};
 
-use crate::{Backtrace, Bytes, ErrorDiagnostic, ext::Extensions};
+use crate::{Backtrace, Bytes, ErrorDiagnostic, ResultType, ext::Extensions};
 
 pub(crate) struct ErrorRepr<E> {
     pub(crate) error: E,
@@ -8,33 +8,6 @@ pub(crate) struct ErrorRepr<E> {
     pub(crate) service: Option<&'static str>,
     pub(crate) backtrace: Backtrace,
     pub(crate) ext: Extensions,
-}
-
-impl<E: ErrorDiagnostic> ErrorRepr<E> {
-    pub(crate) fn new(
-        error: E,
-        service: Option<&'static str>,
-        loc: &'static Location<'static>,
-    ) -> Self {
-        let service = if service.is_none() {
-            error.service()
-        } else {
-            service
-        };
-        let backtrace = if let Some(bt) = error.backtrace() {
-            bt.clone()
-        } else {
-            Backtrace::new(loc)
-        };
-
-        Self {
-            error,
-            service,
-            backtrace,
-            tag: None,
-            ext: Extensions::default(),
-        }
-    }
 }
 
 impl<E: Clone> ErrorRepr<E> {
@@ -123,10 +96,12 @@ impl<E> ErrorDiagnostic for ErrorRepr<E>
 where
     E: ErrorDiagnostic,
 {
-    type Kind = E::Kind;
+    fn typ(&self) -> ResultType {
+        self.error.typ()
+    }
 
-    fn kind(&self) -> Self::Kind {
-        self.error.kind()
+    fn signature(&self) -> &'static str {
+        self.error.signature()
     }
 
     fn tag(&self) -> Option<&Bytes> {
@@ -152,7 +127,7 @@ where
 
 impl<E> error::Error for ErrorRepr<E>
 where
-    E: ErrorDiagnostic,
+    E: ErrorDiagnostic + error::Error,
 {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         Some(&self.error)
@@ -161,7 +136,7 @@ where
 
 impl<E> fmt::Debug for ErrorRepr<E>
 where
-    E: ErrorDiagnostic,
+    E: ErrorDiagnostic + error::Error,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Debug::fmt(&self.error, f)
@@ -170,7 +145,7 @@ where
 
 impl<E> fmt::Display for ErrorRepr<E>
 where
-    E: ErrorDiagnostic,
+    E: ErrorDiagnostic + error::Error,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&self.error, f)